Work on #208 failed to decrypt cipher text after installation and creation of a new account
This problem seems mostly related to cryptographic API usage.

It is not possible to fix the problem without fixing the original issue in the Windows environment.

This pull request adds much more logging to the WinAES class to allow understanding issues.

It also fails the installation if the password encryption could not succeed.